requirements.txt
ciscoconfparse==1.9.41
click==8.1.7
Jinja2==3.1.2
pyats[full]
PyYAML
Requests==2.31.0
ntc_templates
pytest